Add format and codec to entity and changed to constant, moved attributes and timecode to attribute
