+++ ffmpeg-gui/src/main/java/attribute/TimeCode.java
@@ -1,4 +1,4 @@
-package entity;
+package attribute;
 
 import lombok.Getter;
