+++ ffmpeg-gui/src/main/java/constant/AudioCodec.java
@@ -1,4 +1,4 @@
-package entity;
+package constant;
 
 public class AudioCodec {
     public static final String[] AUDIO_CODEC = {"flac", "alac", "wavpack", "tta", "wmalossless", "tak", "aac", "mp4als", "mp3"};
