+++ ffmpeg-gui/src/main/java/constant/AudioFormat.java
@@ -1,4 +1,4 @@
-package entity;
+package constant;
 
 public class AudioFormat {
     //for future extension to Audio
