+++ ffmpeg-gui/src/main/java/constant/VideoCodec.java
@@ -1,4 +1,4 @@
-package entity;
+package constant;
 
 public class VideoCodec {
     public static final String[] VIDEO_CODEC = {"av1", "h264", "hevc", "mpeg1video", "mpeg2video", "mpeg4", "vp9"};
