+++ ffmpeg-gui/src/main/java/constant/VideoFormat.java
@@ -1,4 +1,4 @@
-package entity;
+package constant;
 
 public class VideoFormat {
     public static final String[] VIDEO_FORMATS = {"mp4", "mov", "avi", "webm", "mpeg", "m4v"};
