+++ ffmpeg-gui/src/main/java/entity/AudioJob.java
@@ -1,5 +1,6 @@
 package entity;
 
+import attribute.AudioAttributes;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
